move FindEmptySlotsForOneVolume from volume_growth.go to topology.go
