--- weed/topology/topo_test.go
@@ -1,17 +1,0 @@
-package topology
-
-import (
-	"testing"
-)
-
-func TestRemoveDataCenter(t *testing.T) {
-	topo := setup(topologyLayout)
-	topo.UnlinkChildNode(NodeId("dc2"))
-	if topo.GetActiveVolumeCount() != 15 {
-		t.Fail()
-	}
-	topo.UnlinkChildNode(NodeId("dc3"))
-	if topo.GetActiveVolumeCount() != 12 {
-		t.Fail()
-	}
-}
